+++ src/datasets/Datasets.jl
@@ -1,31 +1,18 @@
-"""
-    module Datasets
-
-Commonly used datasets and utilities for creating data containers.
-
-In the future, contents will be integrated into packages:
-
-- FastAI datasets and data containers will be moved into MLDatasets.jl
-- data container transformations will be moved to MLDataPattern.jl
-
-This submodule will then reexport the same definitions.
-"""
 module Datasets
-
 
 using ..FastAI
 using ..FastAI: typify
 
+import MLUtils: MLUtils, getobs, numobs, filterobs, groupobs, mapobs,
+    shuffleobs, groupobs, eachobs, ObsView
+import MLDatasets
+using MLUtils: mapobs, groupobs
 using DataDeps
-using Glob
 using FilePathsBase
 import DataAugmentation
 using FilePathsBase: filename
 import FileIO
 using IndirectArrays: IndirectArray
-using MLDataPattern
-using MLDataPattern: splitobs
-import LearnBase
 using Colors
 using FixedPointNumbers
 using DataFrames
@@ -36,35 +23,23 @@ using InlineTest
 
 include("fastaidatasets.jl")
 
-function __init__()
-    initdatadeps()
-end
-
+include("batching.jl")
 include("containers.jl")
-include("transformations.jl")
-
 include("load.jl")
+include("loaders.jl")
 include("recipe.jl")
 
 include("deprecations.jl")
 
-include("loaders.jl")
-
+function __init__()
+    initdatadeps()
+end
 
 export
-    # reexports from MLDataPattern
-    splitobs,
-
-    # container transformations
-    mapobs,
-    filterobs,
-    groupobs,
-    joinobs,
-    eachobs,
-
     # primitive containers
-    FileDataset,
     TableDataset,
+
+    mapobs, eachobs, groupobs, shuffleobs, ObsView,
 
     # utilities
     isimagefile,
@@ -78,7 +53,6 @@ export
     grandparentname,
 
     # datasets
-    #DATASETS,
     loadfolderdata,
     datasetpath,